new functionality:
first add a new column 'email' as character to table address

if in table address a record has an entry in the email field (=valid email address) an email will be sent to the mail address specified.
Purpose: Oranisations who donate their data will be informed when one of their records is manipulated by a user.